Remove testing code for IMX708.
Rebuild SO in next branch of libcamera. This solves the ov5647 color changing issue. Due to multiple FOVs available in the pi cam v3 I don't believe there is a good way to support the camera with predefined FOVs since there is no way to tell the difference by the sensor type.