Add ability to lower filesize and import mrcal cli settings
Photonvision only allows importing calibrations that are within a
certain size limit. Currently that is 50000000 bytes, though even
when getting close to that limit, the webUI struggles to import the
files.

To mitigate this, a function was added to strip out observations
from the json to keep the filesize within a specified limit.

A function was added to import mrcal command line intrinsic values
from a camera model file into photonvision json.