More robust network management
This commit changes the way that photonvision interacts with nmcli
to control networking on the coprocessor. Instead of modifying an
existing connection, Photonvision adds new connections for DHCP and
Static IP configurations. It then activiates the proper one at
startup and any time that the network configuration is changed. It
also now uses the interface name and not the connection name and
it checks that the interface is available before making any changes.
If the saved interface is not found, it updates the stored interface
name and applies the network settings to the current interface. This
should minimize the failure to control the network if the network
interface wasn't available when PhotonVision first booted.