First pass at checking if device connected yet
I'm not a fan of this approach yet. It means we have to add a lot of checks and guards around every place we interact with cscore, which kinda blows up combonitorially since this leaks into setting -any- property which happens when we set the current pipeline during VisionModule construction.