More broken refactoring
I'm not sure how to handle this. We really want to delay -all-
processing until the camera itself is connected. And having VisionRunner
do this makes a lot of sense to me. In general i think we need to sort
out our thread model. there's a lot of stuff that happens in the
visionrunner thread vs in a UI HTTP thread vs websockets vs some random
data change thread vs ...